arreglos en la traducción a español, el desert ya se hace tp

+++ Projectiles/DesertTyphoon.cs
@@ -12,55 +12,40 @@ using Terraria.GameContent;
 namespace opswordsII.Projectiles
 {
 
-    public class DesertTyphoon : ModProjectile
-    {
-        public override void SetStaticDefaults()
-        {
-            DisplayName.SetDefault("DesertTyphoon"); //projectile name
-            
-        }
-            public override void SetDefaults()   
-        {    
-            Projectile.width = 36;       //projectile width
-            Projectile.height = 36;  //projectile height
-            Projectile.friendly = true;      //make that the projectile will not damage you
-           Projectile.DamageType = DamageClass.Magic;          // 
-            Projectile.tileCollide = true;   //make that the projectile will be destroed if it hits the terrain
-            Projectile.penetrate = 5;      //how many NPC will penetrate
-            Projectile.timeLeft = 200;   //how many time this projectile has before disepire
-            Projectile.light = 0.75f;    // projectile light
-            Projectile.extraUpdates = 1;
-			Main.projFrames[Projectile.type] = 3;
-            Projectile.ignoreWater = true; 
-            Projectile.aiStyle = ProjectileID.Typhoon;
-			Projectile.CloneDefaults(ProjectileID.DemonScythe);
+	public class DesertTyphoon : ModProjectile
+	{
+		public override void SetStaticDefaults()
+		{
+			DisplayName.SetDefault("DesertTyphoon"); 
 
-        }
+		}
+		public override void SetDefaults()
+		{
+			Projectile.width = 36;
+			Projectile.height = 36;
+			Projectile.friendly = true;
+			Projectile.DamageType = DamageClass.Magic;
+			Projectile.tileCollide = true;
+			Projectile.penetrate = 5;
+			Projectile.timeLeft = 200;
+			Projectile.light = 0.75f;
+			Projectile.extraUpdates = 1;
+			Main.projFrames[Projectile.type] = 3;
+			Projectile.ignoreWater = true;
+			Projectile.aiStyle = ProjectileID.Typhoon;
+			Projectile.CloneDefaults(ProjectileID.DemonScythe);
+		}
 		public override bool OnTileCollide(Vector2 oldVelocity)
 		{
-		Projectile.penetrate--;
-			if (Projectile.penetrate <= 30)
-			{
-				Projectile.Kill();
-			}
+			Projectile.penetrate--;
+			if (Projectile.penetrate <= 30) Projectile.Kill();
 			else
 			{
 				Collision.HitTiles(Projectile.position, Projectile.velocity, Projectile.width, Projectile.height);
 				SoundEngine.PlaySound(SoundID.Item10, Projectile.position);
-
-				// If the projectile hits the left or right side of the tile, reverse the X velocity
-				if (Math.Abs(Projectile.velocity.X - oldVelocity.X) > float.Epsilon)
-				{
-					Projectile.velocity.X = -oldVelocity.X;
-				}
-
-				// If the projectile hits the top or bottom side of the tile, reverse the Y velocity
-				if (Math.Abs(Projectile.velocity.Y - oldVelocity.Y) > float.Epsilon)
-				{
-					Projectile.velocity.Y = -oldVelocity.Y;
-				}
+				if (Math.Abs(Projectile.velocity.X - oldVelocity.X) > float.Epsilon) Projectile.velocity.X = -oldVelocity.X;
+				if (Math.Abs(Projectile.velocity.Y - oldVelocity.Y) > float.Epsilon) Projectile.velocity.Y = -oldVelocity.Y;
 			}
-
 			return false;
 		}
 
@@ -68,8 +53,6 @@ namespace opswordsII.Projectiles
 		{
 			Main.instance.LoadProjectile(Projectile.type);
 			Texture2D texture = TextureAssets.Projectile[Projectile.type].Value;
-
-			// Redraw the projectile with the color not influenced by light
 			Vector2 drawOrigin = new Vector2(texture.Width * 0.5f, Projectile.height * 0.5f);
 			for (int k = 0; k < Projectile.oldPos.Length; k++)
 			{
@@ -82,15 +65,13 @@ namespace opswordsII.Projectiles
 		}
 		public override void AI()
 		{
-			 Projectile.rotation += (float)Projectile.direction * 0.8f;
-			 
+			Projectile.rotation += (float)Projectile.direction * 0.8f;
+
 			if (Projectile.alpha > 70)
 			{
 				Projectile.alpha -= 15;
-				if (Projectile.alpha < 70)
-				{
-					Projectile.alpha = 70;
-				}
+				if (Projectile.alpha < 70) Projectile.alpha = 70;
+
 			}
 			if (Projectile.localAI[0] == 0f)
 			{
@@ -126,61 +107,58 @@ namespace opswordsII.Projectiles
 				}
 			}
 		}
-            private void AdjustMagnitude(ref Vector2 vector)
-            {
-				float magnitude = (float)Math.Sqrt(vector.X * vector.X + vector.Y * vector.Y);
-				if (magnitude > 106f) {
-					vector *= 26f / magnitude;
-				}
-			}
-		
-                   public override void OnHitNPC(NPC target, int damage, float knockback, bool crit) {
-			if (Main.rand.NextBool()) {
-				target.AddBuff(BuffType<Buffs.Burning_Sand>(), 300);
-			}
-        }
-    }
-	
-	 public class DesertTyphoonE : ModProjectile
-    {
-        public override void SetStaticDefaults()
-        {
-            DisplayName.SetDefault("DesertTyphoonE"); //projectile name
-            Main.projFrames[Projectile.type] = 3;
-        }
-            public override void SetDefaults()   
-        {   
-			
-            Projectile.width = 36;       //projectile width
-            Projectile.height = 36;  //projectile height
-            Projectile.friendly = false;      //make that the projectile will not damage you
-           Projectile.DamageType = DamageClass.Melee;   
-			Projectile.hostile =true;       // 
-            Projectile.tileCollide = true;   //make that the projectile will be destroed if it hits the terrain
-            Projectile.penetrate = 5;      //how many NPC will penetrate
-            Projectile.timeLeft = 200;   //how many time this projectile has before disepire
-            Projectile.light = 0.75f;    // projectile light
-            Projectile.extraUpdates = 1;
-            Projectile.ignoreWater = true; 
-            
-			
-			
+		private void AdjustMagnitude(ref Vector2 vector)
+		{
+			float magnitude = (float)Math.Sqrt(vector.X * vector.X + vector.Y * vector.Y);
+			if (magnitude > 106f) vector *= 26f / magnitude;
+		}
 
-        }
-		public override bool OnTileCollide(Vector2 oldVelocity) {
-			//If collide with tile, reduce the penetrate.
-			//So the projectile can reflect at most 5 times
+		public override void OnHitNPC(NPC target, int damage, float knockback, bool crit)
+		{
+			if (Main.rand.NextBool()) target.AddBuff(BuffType<Buffs.Burning_Sand>(), 300);
+
+		}
+	}
+
+	public class DesertTyphoonE : ModProjectile
+	{
+		public override void SetStaticDefaults()
+		{
+			DisplayName.SetDefault("DesertTyphoonE"); 
+			Main.projFrames[Projectile.type] = 3;
+		}
+		public override void SetDefaults()
+		{
+			Projectile.width = 36;      
+			Projectile.height = 36; 
+			Projectile.friendly = false;    
+			Projectile.DamageType = DamageClass.Melee;
+			Projectile.hostile = true;     
+			Projectile.tileCollide = true;   
+			Projectile.penetrate = 5;    
+			Projectile.timeLeft = 200;  
+			Projectile.light = 0.75f;    
+			Projectile.extraUpdates = 1;
+			Projectile.ignoreWater = true;
+		}
+		public override bool OnTileCollide(Vector2 oldVelocity)
+		{
+			
 			Projectile.penetrate--;
-			if (Projectile.penetrate <= 0) {
+			if (Projectile.penetrate <= 0)
+			{
 				Projectile.Kill();
 			}
-			else {
+			else
+			{
 				Collision.HitTiles(Projectile.position + Projectile.velocity, Projectile.velocity, Projectile.width, Projectile.height);
 				SoundEngine.PlaySound(SoundID.Item10, Projectile.position);
-				if (Projectile.velocity.X != oldVelocity.X) {
+				if (Projectile.velocity.X != oldVelocity.X)
+				{
 					Projectile.velocity.X = -oldVelocity.X;
 				}
-				if (Projectile.velocity.Y != oldVelocity.Y) {
+				if (Projectile.velocity.Y != oldVelocity.Y)
+				{
 					Projectile.velocity.Y = -oldVelocity.Y;
 				}
 			}
@@ -205,7 +183,7 @@ namespace opswordsII.Projectiles
 		public override void AI()
 		{
 
-			 Projectile.rotation += (float)Projectile.direction * 0.8f;
+			Projectile.rotation += (float)Projectile.direction * 0.8f;
 			if (Projectile.alpha > 70)
 			{
 				Projectile.alpha -= 15;
@@ -248,18 +226,21 @@ namespace opswordsII.Projectiles
 				}
 			}
 		}
-            private void AdjustMagnitude(ref Vector2 vector)
-            {
-				float magnitude = (float)Math.Sqrt(vector.X * vector.X + vector.Y * vector.Y);
-				if (magnitude > 106f) {
-					vector *= 26f / magnitude;
-				}
+		private void AdjustMagnitude(ref Vector2 vector)
+		{
+			float magnitude = (float)Math.Sqrt(vector.X * vector.X + vector.Y * vector.Y);
+			if (magnitude > 106f)
+			{
+				vector *= 26f / magnitude;
 			}
-		
-                   public override void OnHitNPC(NPC target, int damage, float knockback, bool crit) {
-			if (Main.rand.NextBool()) {
+		}
+
+		public override void OnHitNPC(NPC target, int damage, float knockback, bool crit)
+		{
+			if (Main.rand.NextBool())
+			{
 				target.AddBuff(BuffType<Buffs.Burning_Sand>(), 300);
 			}
-        }
-  }
-  }
+		}
+	}
+}
